make CodegenLayout internal again, use CodegenMetadata.resolveSchemaType instead
