+++ libraries/apollo-compiler/src/main/kotlin/com/apollographql/apollo3/compiler/CodegenMetadata.kt
@@ -1,5 +1,7 @@
 package com.apollographql.apollo3.compiler
 
+import com.apollographql.apollo3.annotations.ApolloInternal
+import com.apollographql.apollo3.compiler.codegen.ResolverClassName
 import com.apollographql.apollo3.compiler.codegen.ResolverInfo
 import com.apollographql.apollo3.compiler.codegen.ResolverKeyKind
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -22,6 +24,11 @@ data class CodegenMetadata internal constructor(
 
 fun CodegenMetadata.schemaTypes(): Set<String> {
   return resolverInfo.entries.filter { it.key.kind == ResolverKeyKind.SchemaType }.map { it.key.id }.toSet()
+}
+
+@ApolloInternal
+fun CodegenMetadata.resolveSchemaType(name: String): ResolverClassName? {
+  return resolverInfo.entries.firstOrNull { it.key.kind == ResolverKeyKind.SchemaType && it.key.id == name }?.className
 }
 
 @OptIn(ExperimentalSerializationApi::class)
