+++ libraries/apollo-compiler/src/main/kotlin/com/apollographql/apollo3/compiler/CodegenSchema.kt
@@ -1,6 +1,5 @@
 package com.apollographql.apollo3.compiler
 
-import com.apollographql.apollo3.annotations.ApolloInternal
 import com.apollographql.apollo3.ast.Schema
 import com.apollographql.apollo3.ast.findTargetName
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -28,14 +27,12 @@ class CodegenSchema(
     val generateDataBuilders: Boolean,
 )
 
-@ApolloInternal
-class CodegenType(
+internal class CodegenType(
     val name: String,
     val targetName: String?,
 )
 
-@ApolloInternal
-fun CodegenSchema.allTypes(): List<CodegenType> {
+internal fun CodegenSchema.allTypes(): List<CodegenType> {
   return schema.typeDefinitions.values.map {
     CodegenType(it.name, it.directives.findTargetName(schema))
   }.sortedBy { it.name }
